backend/requirements.txt
fastapi
uvicorn[standard]
python-dotenv
tenacity
google-generativeai
google-cloud-storage
firebase-admin
pandas
streamlit
plotly
matplotlib
seaborn
psutil
pygwalker
hiplot
streamlit-hub